+++ src/test/testUtils.ts
@@ -3,13 +3,12 @@ import * as io from '../io';
 import * as models from '../models';
 import '../registerPlugins';
 import * as store from '../store';
-import { promises as fs } from 'fs';
 import { EOL } from 'os';
 import { isAbsolute, dirname, extname, join } from 'path';
 
-export async function parseHttp(code: string) {
-  return await new store.HttpFileStore().getOrCreate(`any.http`, async () => Promise.resolve(code), 0, {
-    workingDir: __dirname,
+export async function parseHttp(code: string, filename = 'any.http') {
+  return await new store.HttpFileStore().getOrCreate(filename, async () => Promise.resolve(code), 0, {
+    workingDir: '/',
   });
 }
 
@@ -45,7 +44,7 @@ export function initFileProvider(files?: Record<string, string> | undefined) {
   const fileProvider = io.fileProvider;
   fileProvider.EOL = EOL;
 
-  fileProvider.isAbsolute = async fileName => isAbsolute(fileProvider.toString(fileName));
+  fileProvider.isAbsolute = async fileName => typeof fileName === 'string' && fileName.startsWith('/');
   fileProvider.dirname = fileName => dirname(fileProvider.toString(fileName));
   fileProvider.hasExtension = (fileName, ...extensions) =>
     extensions.indexOf(extname(fileProvider.toString(fileName))) >= 0;
@@ -63,10 +62,72 @@ export function initFileProvider(files?: Record<string, string> | undefined) {
     }
     throw new Error('No File');
   };
-  fileProvider.readdir = async dirname => {
+  fileProvider.readdir = async () => {
     if (files) {
       return Object.keys(files);
     }
-    return fs.readdir(fileProvider.toString(dirname));
+    return [];
   };
+}
+
+export type Directory = {
+  childs: Record<string, Directory | string>;
+};
+
+export function initNestedFileProvider(filesystem: Directory) {
+  const fileProvider = io.fileProvider;
+  fileProvider.EOL = EOL;
+
+  fileProvider.isAbsolute = async fileName => isAbsolute(fileProvider.toString(fileName));
+  fileProvider.dirname = fileName => dirname(fileProvider.toString(fileName));
+  fileProvider.hasExtension = (fileName, ...extensions) =>
+    extensions.indexOf(extname(fileProvider.toString(fileName))) >= 0;
+  fileProvider.joinPath = (fileName, path) => join(fileProvider.toString(fileName), path);
+
+  function getPathContent(file: string): Directory | string | undefined {
+    const paths = file.split('/').slice(1).reverse();
+    let path: string | undefined;
+    let currentObj: Directory | string | undefined = filesystem;
+    while ((path = paths.pop()) && currentObj) {
+      if (isDirectory(currentObj)) {
+        currentObj = currentObj.childs[path];
+      } else {
+        return undefined;
+      }
+    }
+    return currentObj;
+  }
+
+  fileProvider.exists = async fileName => Promise.resolve(typeof fileName === 'string' && !!getPathContent(fileName));
+  fileProvider.readFile = async filename => {
+    if (typeof filename === 'string') {
+      const content = getPathContent(filename);
+      if (typeof content === 'string') {
+        return content;
+      }
+    }
+    throw new Error('No File');
+  };
+  fileProvider.readBuffer = async filename => {
+    if (typeof filename === 'string') {
+      const content = getPathContent(filename);
+      if (typeof content === 'string') {
+        return Buffer.from(content);
+      }
+    }
+    throw new Error('No File');
+  };
+  fileProvider.readdir = async dirname => {
+    if (typeof dirname === 'string') {
+      const content = getPathContent(dirname);
+      if (isDirectory(content)) {
+        return Object.keys(content.childs);
+      }
+    }
+    return [];
+  };
+}
+
+function isDirectory(file: Directory | string | undefined): file is Directory {
+  return file !== undefined && typeof file !== 'string';
 }